+++ demo/CHANGELOG.md
@@ -1,5 +1,25 @@
 # demo
 
+## 1.0.57
+
+### Patch Changes
+
+- Updated dependencies [c12d135]
+  - @knime/virtual-tree@1.2.0
+  - @knime/components@1.4.5
+  - @knime/rich-text-editor@1.2.29
+
+## 1.0.56
+
+### Patch Changes
+
+- Updated dependencies [4230d50]
+- Updated dependencies [e5fff5a]
+- Updated dependencies [9364d3a]
+  - @knime/virtual-tree@1.1.0
+  - @knime/components@1.4.4
+  - @knime/rich-text-editor@1.2.28
+
 ## 1.0.55
 
 ### Patch Changes